👽️ Middleware >> Made a few functions into middlewares for better and cleaner access.
